BIOS-1244
PROBLEM:
Polling in agent doesn't work as expected. If messages keep coming
before the
end of timeout period, onPoll () is not executed when it should.

PROBLEM:
Code design is not logical, hard to understand and it makes it hard to
extend it.

PROBLEM:
In general it is not possible to take a different action in
NUTConfigurator
regarding rules than sending them out each time any kind of asset
operation is
performed (i.e. it is not possible to not do anything, or to perform
theoretical delete).

PROBLEM:
If for some reason sendNewRules () fail, the device is flagged as
configured
anyway and agent-autoconfig will not re-try configuration of rules for
epdu/ups.

PROBLEM:
When as asset message with update asset operation arrives, configuration
never
takes place due to the way addDeviceIfNeeded works. This is true even
for two
consecutive insert messages.

PROBLEM:
For every asset to be configure a new / delete memory allocation is
performed
(i.e. even if there are thousands of assets to be configured by the same
Configurator consecutively, we will do thousand new/delete's
consecutively),
even though the configuration operation is state-less and the program
execution
is single-threaded (synchronous).

PROBLEM:
Configuration of uptime agent is done in hacky way. Duplicated
malamute client is being created, the trigger bypasses autoconfig
mechanism...

JOINT SOLUTION:
total rewamp :)

Signed-off-by: Karol Hrdina <karolhrdina@eaton.com>