Allow to generate the release-details files in the running system
To untangle the initramfs from the booted image a bit, add support for
running generate-release-details.sh in the running system, using
variables exported by the initramfs. ARMv7 stays unchanged, because the
infrastructure (/run/initramfs) is missing, but on x86_64 with dracut,
we can reduce the number of hacks and make the initramfs and the image
more independent from each other.