Revert "fty-proxy.service.in : cause the service to try-restart (if running) its consumers upon (re-)enablement"
This breaks the EULA acceptance, because the restart command eventually
kills the start-db-services script. This revert is a minimal change to
unbreak EULA acceptance. I am aware that it breaks proxy handling in
return, but this should be properly solved by making users of the proxy
re-read the file each time time an http request is made, instead of
restarting processes.

This reverts commit e19908cc415632734518cf5bdd4a182c3af55a51.