Implement runas-_bios-script utility
This utility allows tntnet to spawn commands as _bios-script. This is
intended for running user scripts provided by fty-scripts-rest and
maybe other components in the future.

Precautions were taken to prevent string injection attacks, but this
is a safe design only as far as user priviledges go. A more thorough
sandboxing scheme would require the usage of Linux namespaces.